Implementing list of Spaces
