+++ ConfluenceEX/Main/ConfluenceToolWindowNavigator.xaml
@@ -9,8 +9,8 @@
              d:DesignHeight="300" d:DesignWidth="300">
 
     <UserControl.Resources>
-        <DataTemplate DataType="{x:Type local:ContentListView}">
-            <local:ContentListView/>
+        <DataTemplate DataType="{x:Type local:SpaceListView}">
+            <local:SpaceListView/>
         </DataTemplate>
         <DataTemplate DataType="{x:Type local:BeforeSignInView}">
             <local:BeforeSignInView/>
