+++ ConfluenceEX/View/ContentView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ConfluenceEX.View.ContentListView"
+﻿<UserControl x:Class="ConfluenceEX.View.ContentView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -18,32 +18,9 @@
             <Setter Property="Foreground" Value="{DynamicResource {x:Static vsp:EnvironmentColors.SystemMenuTextBrushKey}}"/>
         </Style>
     </UserControl.Resources>
-        
+    
     <Grid Margin="10">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto" />
-            <ColumnDefinition Width="Auto" />
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="Auto" />
-        </Grid.RowDefinitions>
-
-        <Label Grid.Column="0" Grid.Row="0">Id:</Label>
-        <Label Grid.Column="0" Grid.Row="1">Type:</Label>
-        <Label Grid.Column="0" Grid.Row="2">Status:</Label>
-        <Label Grid.Column="0" Grid.Row="3">Title:</Label>
-        <TextBlock Grid.Column="1" Grid.Row="0" Text="{Binding Content.Id}"/>
-        <TextBlock Grid.Column="1" Grid.Row="1" Text="{Binding Content.Type}"/>
-        <TextBlock Grid.Column="1" Grid.Row="2" Text="{Binding Content.Status}"/>
-        <TextBlock Grid.Column="1" Grid.Row="3" Text="{Binding Title}"/>
-
-        <ListBox Grid.Column="0" Grid.Row="4" ItemsSource="{Binding ContentList}"
+        <ListBox Grid.Column="0" Grid.Row="0" ItemsSource="{Binding SpaceContentList}"
                  DisplayMemberPath="Title"/>
     </Grid>
-   
 </UserControl>
-
