+++ ConfluenceEX/View/SpaceListView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ConfluenceEX.View.SpaceContentView"
+﻿<UserControl x:Class="ConfluenceEX.View.SpaceListView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -18,9 +18,20 @@
             <Setter Property="Foreground" Value="{DynamicResource {x:Static vsp:EnvironmentColors.SystemMenuTextBrushKey}}"/>
         </Style>
     </UserControl.Resources>
-    
+        
     <Grid Margin="10">
-        <ListBox Grid.Column="0" Grid.Row="0" ItemsSource="{Binding SpaceContentList}"
-                 DisplayMemberPath="Title"/>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+        </Grid.ColumnDefinitions>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+
+        <Label Grid.Column="0" Grid.Row="0">Spaces list:</Label>
+        <ListView Grid.Column="0" Grid.Row="1" ItemsSource="{Binding SpaceList}"
+                 DisplayMemberPath="Name"/>
     </Grid>
+   
 </UserControl>
+
